WATER1 + WATER-AUDIT: the enhanced water surface over the oceans, rivers and ponds, audited

+++ src/formats/umRandom.js
@@ -50,9 +50,13 @@ export class UMRandom {
     return F32[0] - 1.0;
   }
 
-  /** Uniform float in [min, max). */
+  /** Uniform float in [min, max). WATER-AUDIT (2026-09-08): float32 per
+   *  operation, as `NextFloat() * (max - min) + min` is in C# - the
+   *  double form disagreed with the reference in 17% of draws by a
+   *  float32 ulp, and through the beach jitter's `fround(40 + j)` moved
+   *  the tile threshold itself for 1% of corners. */
   nextFloatRange(min, max) {
-    return this.nextFloat() * (max - min) + min;
+    return Math.fround(Math.fround(this.nextFloat() * Math.fround(max - min)) + min);
   }
 
   /** Uniform int in [min, max), verbatim: (state * range) >> 32 + min. */
